+++ tests/tolerances/test_all_objective_results_ok.py
@@ -26,7 +26,9 @@ def test_all_objective_results_ok_doesnt_log_table_when_results_all_ok(
     mocked_logger: MagicMock, objective_results_all_ok: List[ObjectiveResult]
 ) -> None:
     _ = all_objective_results_ok(objective_results_all_ok)
-    mocked_logger.info.assert_called_once_with("All Objective Results were OK.")
+    mocked_logger.debug.assert_called_once_with(
+        "All Objective Results were OK."
+    )
 
 
 @patch("chaosreliably.slo.tolerances.logger")